+++ libbpf-rs/CHANGELOG.md
@@ -3,6 +3,7 @@ Unreleased
 - Added `KprobeMultiLinkInfo` & `UprobeMultiLinkInfo` fields that required a
   second info call.
 - Added `key_size`, `value_size` and more getters to `OpenMap{,Mut}`
+- Added specific `IterLinkInfo` fields for `LinkTypeInfo::Iter`.
 
 
 0.26.2